pyproject.toml
[project]
name = "did-web-annuaire-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django==5.2.11",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-htmx>=1.27.0",
    "django-ninja>=1.5.3",
    "django-ninja-extra>=0.31.0",
    "django-ninja-jwt>=5.4.4",
    "django-structlog>=10.0.0",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "rav>=0.4.3",
    "redis>=7.2.0",
    "structlog>=25.5.0",
]
